Split up gateway.rs into several files
